Move glibui and libui into spikes
These feel still in the spike stage and I want to make that clear. Yes,
having them in this directory means they won't necessarily run now, but
we're at the point where I don't want to cause confusion for folks
navigating in lib/scarpe.

Going forward, we can move display services back in, but I think we'd
want to nominate and have a "responsible person" for each display
service, in the README.

I still find this code valuable as artifacts and am very happy to keep
it in spikes.

**For Reviewers**

Is there context I'm missing here? Have these developed and should we
keep them in this directory?