Do not force Bloops on people unless they use it
This also introduces a concept of unforced dependencies. Yes,
Shoes *comes* with everything. But I don't think folks running
shoes should stumble across the loading issues *unless* they actually
execute the thing.

So the idea is "unforced dependencies". We require them in when
and where you use them. But for now just Bloops, as it's been an
incredible headwind for many engineers trying to get started.