Various fixes related to getting the ovals demo working better. (#511)
Fix and test stroke and fill for art drawables.
Fix absolute positioning to be relative to nearest enclosing slot.
Fix stroke and fill to convert colors to hex-RGB properly.
Fix span to not use html_attributes, which it was mostly doing incorrectly anyway.

Note that a number of the fixtures changes are to switch back and forth between black and #000000.
That's not great, and it's an artifact of how we default colors in several places.
Once we have colors in a component I think we'll want to always output RGB hex format,
which will take care of that problem. In the mean time, I'm keeping a lot of code that's
already in place.